Convert into a two module gradle project: json-proxy-core is the thin component artifact and json-proxy-service is the fat standalone server
