+++ .gitignore
@@ -14,3 +14,8 @@ hs_err_pid*
 .classpath
 .project
 .settings
+.gradle
+build/
+.idea/
+*.swp
+*.iml
