--- Application.groovy
@@ -1,6 +1,0 @@
-@Grab("edu.wisc.my.restproxy:json-proxy-service:1.1.6-SNAPSHOT")
-
-@ComponentScan("edu.wisc.my.restproxy")
-@RestController
-class ProxyService {}
-
