--- application.yml
@@ -1,53 +1,0 @@
-grails:
-    profile: web-micro
-    codegen:
-        defaultPackage: proxy.service
-info:
-    app:
-        name: proxy-service
-        version: 0.1
-        grailsVersion: 3.0.7
-spring:
-    groovy:
-        template:
-            check-template-location: false
----
-grails:
-    mime:
-        disable:
-            accept:
-                header:
-                    userAgents:
-                        - Gecko
-                        - WebKit
-                        - Presto
-                        - Trident
-        types:
-            all: '*/*'
-            atom: application/atom+xml
-            css: text/css
-            csv: text/csv
-            form: application/x-www-form-urlencoded
-            html: 
-              - text/html
-              - application/xhtml+xml
-            js: text/javascript
-            json:
-              - application/json
-              - text/json
-            multipartForm: multipart/form-data
-            rss: application/rss+xml
-            text: text/plain
-            hal: 
-              - application/hal+json
-              - application/hal+xml
-            xml:
-              - text/xml
-              - application/xml
-    urlmapping:
-        cache:
-            maxsize: 1000
-    controllers:
-        defaultScope: singleton
-    converters:
-        encoding: UTF-8
