+++ build.gradle
@@ -1,0 +1,15 @@
+subprojects {
+
+    group = 'edu.wisc.my.restproxy'
+    version = '1.2.0-SNAPSHOT'
+
+    apply plugin: 'groovy'
+
+    repositories {
+        maven { url "https://oss.sonatype.org/content/repositories/snapshots/" }
+        maven { url "https://code.doit.wisc.edu/maven/content/repositories/public-releases/" }
+        maven { url "https://code.doit.wisc.edu/maven/content/repositories/uw-releases/" }
+        maven { url "http://repo.maven.apache.org/maven2" }
+    }
+
+}
