--- endpoint.properties
@@ -1,28 +1,0 @@
-#key.uri                : URI for service with {attr} for attributes
-#key.username           : the username to login to service (basic auth) *
-#key.password           : the password for the service (basic auth)     *
-#key.attributes         : the csv of attributes
-#key.cache              : how long to cache the information for that URI (with variables replaced) - not implemented yet
-#
-# * : not required
-
-#Wiscmail lookup
-wiscmaillookup.uri=https://email-server.wisc.edu/some.json?action=listaccountInfo&account={uid}
-wiscmaillookup.username=
-wiscmaillookup.password=
-wiscmaillookup.attributes=uid
-
-#new stuff feed
-newstuff.uri=https://example.com/feed.json?username={uid}&name={prefname}
-newstuff.username=
-newstuff.password=
-newstuff.attributes=uid,prefname
-
-#subreddit feed example
-uwmadisonreddit.uri=https://www.reddit.com/r/uwmadison/hot.json
-uwmadisonreddit.username=
-uwmadisonreddit.password=
-uwmadisonreddit.attributes=
-
-#todos placeholder example
-todos.uri=http://jsonplaceholder.typicode.com/todos
