+++ json-proxy-core/src/main/groovy/edu/wisc/my/restproxy/RequestBody.groovy
