+++ json-proxy-core/src/main/groovy/edu/wisc/my/restproxy/api/GenericRestLookupController.groovy
