+++ json-proxy-core/src/main/groovy/edu/wisc/my/restproxy/dao/GenericRestLookupDao.groovy
