+++ json-proxy-core/src/main/groovy/edu/wisc/my/restproxy/dao/GenericRestLookupDaoImpl.groovy
