+++ json-proxy-core/src/main/groovy/edu/wisc/my/restproxy/service/RestProxyService.groovy
