+++ json-proxy-core/src/main/groovy/edu/wisc/my/restproxy/web/ResourceProxyController.groovy
