+++ json-proxy-core/src/main/resources/.gitignore
