+++ json-proxy-core/src/main/resources/endpoint.properties.example
