+++ json-proxy-service/src/main/groovy/edu/wisc/my/restproxy/Server.groovy
@@ -1,0 +1,13 @@
+package edu.wisc.my.restproxy
+
+import org.springframework.boot.SpringApplication
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.autoconfigure.*
+
+@SpringBootApplication
+class Server {
+    public static void main(String[] args) {
+        SpringApplication.run(Server.class, args);
+    }
+}
+
