+++ json-proxy-service/src/main/resources/.gitignore
@@ -1,0 +1,2 @@
+endpoint.properties
+
