--- src/test/java/edu/wisc/my/restproxy/ValidationResult.java
@@ -1,23 +1,0 @@
-/**
- * 
- */
-package edu.wisc.my.restproxy;
-
-/**
- * Just an example of an object that could be retruned by rest api.
- * @author Collin Cudd
- */
-public class ValidationResult {
-
-  boolean success;
-  String message;
-  /**
-   * @param success
-   * @param message
-   */
-  public ValidationResult(boolean success, String message) {
-    this.success = success;
-    this.message = message;
-  }
-  
-}
